load emb update

+++ Main_embedding.py
@@ -17,28 +17,32 @@ def main(label, cls=MLA, exp=True):
     print(f'\nLoading {model_name} Base Model...')
     basemodel = BaseModel(model_name)
     model = basemodel.load_model(True)
-    # if model_name == 'vggface':
-    #     model = basemodel.adding_toplayer(model)
 
-    # todo add file embedding lookout
-    # emb_file = 'data_'+model_name+'_'+label'.csv'
-    # if os.path.exist(os.path.join(EMB_PATH, emb_file))
-    # train, test = read_file[x], [y]
-
-    # Start images processing and dataframe splitting
+    # Save or Load Embedding
+    emb_file = 'data_' + model_name + '_' + label + '.csv'
+    emb_path = os.path.join(EMB_PATH, emb_file)
     print('Reading File...\nCreating Train, Test...')
-    trainer = Train(IND_FILE, IMAGE_PATH)
     print(f'\nTrain on {label} attribute\n')
-    train, test = trainer.data_preprocess(IND_FILE, label, 50, True, None)
-    print('Done!')
 
-    # Save embedding
-    print(f'\nSave Embedding...')
-    X_train, y_train = basemodel.loading_embedding(IMAGE_PATH, model, train, 1)
-    X_test, y_test = basemodel.loading_embedding(IMAGE_PATH, model, test, 1)
-    data_emb = pd.DataFrame(np.vstack([X_train, X_test]))
-    data_emb.to_csv(os.getcwd()+'/csv/data/data_'+model_name+'_'+label+'.csv')
+    if os.path.exists(emb_path):
+        print('Load Saved Embedding...')
+        df_temp = pd.read_csv(emb_path)
+        X_train, X_test = df_temp.iloc[:0.8*len(df), 0], df_temp.iloc[0.8*len(df):, 1]
+        y_train, y_test = df_temp.iloc[:0.8*len(df), 2], df_temp.iloc[0.8*len(df):, 3]
+    else:
+        # Start images processing and dataframe splitting
+        trainer = Train(IND_FILE, IMAGE_PATH)
+        train, test = trainer.data_preprocess(IND_FILE, label, 50, True, None)
+        print('Done!')
 
+        # Save embedding
+        print(f'\nSave Embedding...')
+        X_train, y_train = basemodel.loading_embedding(IMAGE_PATH, model, train, 1)
+        X_test, y_test = basemodel.loading_embedding(IMAGE_PATH, model, test, 1)
+        data_emb = pd.DataFrame(np.hstack(np.vstack([X_train, X_test]), np.vstack([y_train, y_test])))
+        data_emb.to_csv(os.getcwd() + '/csv/data/data_' + model_name + '_' + label + '.csv')
+
+    # GridSearch Classifiers
     if not isinstance(cls, str):
         print('GridSearch top Cls...')
         df_cls = gridsearch_cls(X_train, y_train, X_test, y_test, cls)
@@ -75,9 +79,9 @@ def main(label, cls=MLA, exp=True):
     y_pred = df_top_cls['MLA pred'].values[0]
 
     # Saving embedding and final results to file
-    label_emb = pd.DataFrame({'y_test': pd.Series(y_test), 'y_pred': pd.Series(y_pred)})
-    label_emb.to_csv(os.getcwd()+'/csv/data/label_'+model_name+'_'+label+'_'+cls_name+'.csv')
-    df_top_cls.to_csv(os.getcwd()+'/csv/data/sum_'+model_name+'_'+label+'_top3.csv')
+    # label_emb = pd.DataFrame({'y_test': pd.Series(y_test), 'y_pred': pd.Series(y_pred)})
+    # label_emb.to_csv(os.getcwd() + '/csv/data/label_' + model_name + '_' + label + '_' + cls_name + '.csv')
+    df_top_cls.to_csv(os.getcwd() + '/csv/data/sum_' + model_name + '_' + label + '_top3.csv')
 
     # plot confusion matrix and acc score
     if not exp:
@@ -102,28 +106,28 @@ if __name__ == '__main__':
     cols = df.columns.tolist()
     # accessories_label = [l for l in cols if l.startswith("Wearing")]
     labels = ['Arched_Eyebrows',
- 'Attractive',
- 'Bags_Under_Eyes',
- 'Bald',
- 'Bangs',
- 'Big_Lips',
- 'Big_Nose',
- 'Bushy_Eyebrows',
- 'Chubby',
- 'Double_Chin',
- 'Eyeglasses',
- 'Goatee',
- 'Heavy_Makeup',
- 'High_Cheekbones',
- 'Mustache',
- 'Narrow_Eyes',
- 'No_Beard',
- 'Oval_Face',
- 'Pale_Skin',
- 'Pointy_Nose',
- 'Receding_Hairline',
- 'Rosy_Cheeks',
- 'Sideburns',
- 'Smiling']
+              'Attractive',
+              'Bags_Under_Eyes',
+              'Bald',
+              'Bangs',
+              'Big_Lips',
+              'Big_Nose',
+              'Bushy_Eyebrows',
+              'Chubby',
+              'Double_Chin',
+              'Eyeglasses',
+              'Goatee',
+              'Heavy_Makeup',
+              'High_Cheekbones',
+              'Mustache',
+              'Narrow_Eyes',
+              'No_Beard',
+              'Oval_Face',
+              'Pale_Skin',
+              'Pointy_Nose',
+              'Receding_Hairline',
+              'Rosy_Cheeks',
+              'Sideburns',
+              'Smiling']
     for label in labels:
         main(label, MLA, True)
